+++ packages/create-houdini/bin.js
@@ -1,7 +1,8 @@
 #!/usr/bin/env node
 import * as p from '@clack/prompts'
+import { program, Option, InvalidArgumentError } from 'commander'
 import * as graphql from 'graphql'
-import { bold, cyan, gray, green, grey, italic, white } from 'kleur/colors'
+import { bold, cyan, gray, grey, italic, white } from 'kleur/colors'
 import fs, { readFileSync, writeFileSync } from 'node:fs'
 import path from 'node:path'
 import { exit } from 'node:process'
@@ -28,6 +29,24 @@ const options = fs.readdirSync(templatesDir).map((templateDir) => {
 	return { ...data, value: templateDir }
 })
 
+program.argument('[project_name]', 'optional project name')
+program.addOption(
+	new Option('-t, --template <template>', 'template you want to use').choices(
+		options.map((c) => c.value)
+	)
+)
+program.addOption(
+	new Option('-s, --schema <schema>', '"local" or "http..."').argParser((value) => {
+		if (value === 'local' || value.startsWith('http')) {
+			return value
+		}
+		throw new InvalidArgumentError('Should be "local" or "http..." or do not set it!')
+	})
+)
+
+program.parse(process.argv)
+const options_cli = program.opts()
+
 p.intro('🎩 Welcome to Houdini!')
 
 // if we weren't given a directory, then we should ask
@@ -43,7 +62,7 @@ if (!projectDir) {
 
 	if (dir) {
 		projectDir = dir
-		projectName = 'hello-houdini'
+		projectName = dir
 	} else {
 		projectDir = '.'
 	}
@@ -78,11 +97,13 @@ if (dirToCreate && !fs.existsSync(projectDir)) {
 	fs.mkdirSync(projectDir)
 }
 
-const template = await p.select({
-	message: 'Which template do you want to use?',
-	initialValue: 'react-typescript',
-	options,
-})
+const template = options_cli.template
+	? options_cli.template
+	: await p.select({
+			message: 'Which template do you want to use?',
+			initialValue: 'react-typescript',
+			options,
+	  })
 if (p.isCancel(template)) {
 	process.exit(1)
 }
@@ -94,14 +115,18 @@ if (!templateMeta) {
 }
 
 // ask if the schema is local or remote
-const localSchema =
-	template !== 'sveltekit-demo' &&
-	(await p.confirm({
-		message: 'Is your api going to be defined in this project too?',
-	}))
+const localSchema = templateMeta.apiUrl
+	? false
+	: options_cli.schema === 'local'
+	? true
+	: options_cli.schema?.startsWith('http')
+	? false
+	: await p.confirm({
+			message: 'Is your api going to be defined in this project too?',
+	  })
 
 // if we have a remote schema then we need to introspect it and write the value
-let apiUrl = templateMeta.apiUrl ?? ''
+let apiUrl = options_cli.schema?.startsWith('http') ? options_cli.schema : templateMeta.apiUrl ?? ''
 if (!localSchema) {
 	let pullSchema_content = ''
 	if (apiUrl === '') {
@@ -150,16 +175,13 @@ copy(
 		["'CLIENT_CONFIG'"]: clientConfig,
 		["'CONFIG_FILE'"]: configFile,
 	},
+	{ '.meta.gitignore': '.gitignore' },
 	['.meta.json']
 )
 
-// if we have a local schema then we have more fiiles to copy
+// if we have a local schema then we have more files to copy
 if (localSchema) {
-	if (template === 'react') {
-		copy(sourcePath('./fragments/localApi'))
-	} else if (template === 'react-typescript') {
-		copy(sourcePath('./fragments/localApi-typescript'))
-	}
+	copy(sourcePath('./fragments/localSchema/' + template))
 }
 
 // If anything goes wrong, we don't want to block the user
@@ -192,6 +214,7 @@ function copy(
 	/** @type {string} */ sourceDir,
 	/** @type {string} */ destDir = projectDir,
 	/** @type {Record<string, string>} */ transformMap = {},
+	/** @type {Record<string, string>} */ transformFileMap = {},
 	/** @type {string[]} */ ignoreList = []
 ) {
 	if (!fs.existsSync(destDir)) {
@@ -199,12 +222,16 @@ function copy(
 	}
 
 	const files = fs.readdirSync(sourceDir)
-	for (const file of files) {
-		const sourceFilePath = path.join(sourceDir, file)
+	for (const fileSource of files) {
+		const fileDest = Object.entries(transformFileMap).reduce((acc, [key, value]) => {
+			return acc.replace(key, value)
+		}, fileSource)
+		// const file = fileSource.replace(".meta.gitignore", ".gitignore")
+		const sourceFilePath = path.join(sourceDir, fileSource)
 		const sourceRelative = path.relative(templateDir, sourceFilePath)
 		// skip the ignore list
 		if (!ignoreList.includes(sourceRelative)) {
-			const destFilePath = path.join(destDir, file)
+			const destFilePath = path.join(destDir, fileDest)
 
 			const stats = fs.statSync(sourceFilePath)
 
@@ -226,7 +253,7 @@ function copy(
 			}
 			// if we run into a directory then we should keep going
 			else if (stats.isDirectory()) {
-				copy(sourceFilePath, destFilePath, transformMap, ignoreList)
+				copy(sourceFilePath, destFilePath, transformMap, transformFileMap, ignoreList)
 			}
 		}
 	}
@@ -274,7 +301,8 @@ async function pullSchemaCli() {
 
 		url_and_headers = answer.url_and_headers
 		const value_splited = url_and_headers.split(' ')
-		const apiUrl = value_splited[0]
+
+		apiUrl = value_splited[0]
 
 		const local_headers =
 			value_splited.length > 1