+++ packages/houdini-react/src/runtime/server/utils/isCallable.ts
@@ -1,0 +1,5 @@
+export { isCallable }
+
+function isCallable<T extends (...args: unknown[]) => unknown>(thing: T | unknown): thing is T {
+	return thing instanceof Function || typeof thing === 'function'
+}